chore: move queries and utils to new module
